Develop

See merge request s05-webmobile1-sub3/S05P13C103!128

Former-commit-id: 0760a4cf88f5e4cb920b769546f08babf5f0cf86

+++ backend/src/main/java/com/ssafy/api/response/dto/Mypage/MypagePRes.java
@@ -1,7 +1,5 @@
 package com.ssafy.api.response.dto.Mypage;
 
-import com.ssafy.db.entity.User;
-
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
@@ -15,6 +13,6 @@ import lombok.Setter;
 @ApiModel("MypageProductResponse")
 public class MypagePRes {
 	@ApiModelProperty(name="Product PK")
-	Long id;
+	Integer id;
 	String title;
 }
